+++ silence/sql/validator.py
@@ -1,5 +1,0 @@
-import sqlvalidator
-
-def sql_is_ok(sql):
-    parsed = sqlvalidator.parse(sql)
-    return parsed.is_valid()