+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.2.0
+- Added support for HTTP POST/PUT/DELETE and SQL INSERT/UPDATE/DELETE
+- Added support for static web files
+
 # 0.1.0
 - Added support for HTTP GET / SQL SELECT (URL params, query filtering...)